requirements.txt
numpy
pandas
setuptools
mysql-connector-python
scikit-learn
lightgbm
xgboost
tensorboard
torch
mlflow
dvc
